% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssolv.R
\name{bssolv}
\alias{bssolv}
\title{Bounded secant solver for a 1-D function of integer values}
\usage{
bssolv(x1, x2, targfun)
}
\arguments{
\item{x1}{lower bound}

\item{x2}{upper bound}

\item{targfun}{function to solve}
}
\description{
This solver is specialized to the case where the function being solved is
only defined for integer values.  We still pass noninteger values to the
function, but we expect that they will be rounded.  This affects our stopping
criteria.
}
\keyword{internal}
