% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnetfn.R
\name{pnetfn}
\alias{pnetfn}
\title{Calculate the probability of a specified net false negatives}
\usage{
pnetfn(k, kpos, delta, phi, eta)
}
\arguments{
\item{k}{Number of trials}

\item{kpos}{Number of positive samples in the sample pool (before test error
is applied)}

\item{delta}{The number of net false negatives}

\item{phi}{Sensitivity (true positive rate) of the test.}

\item{eta}{Specificity (true negative rate) of the test.}
}
\description{
The net false negatives are the number of false negatives minus the number
of false positives.  It is, in other words, the number of counts by which
the observed positives are shifted downward, relative to the actual number
of positive counts in the sample.  This function calculates the probability
that the net false negatives equal a specified value.
}
\details{
This probability can be calculated in terms of the probability mass function
for the binomial distribution, \eqn{\rho_{bn}} (\code{\link[stats:Binomial]{dbinom}} in
R-lingo).
\deqn{P(NFN = \delta) = \sum_{j=0}^{k} \rho_{bn}(j, k, 1-\phi) \rho_{bn}(j-\delta, k, 1-\eta),}
where \eqn{\phi} is the sensitivity of the test and \eqn{\eta} is the specificity.

Note that this function is not vectorized.
}
